Removed Extender as it is not flexible to add new themes. Added the ITheme derived classes to manage theme elements.
